Move OnDiskContent class
