+++ Mnema.API/Content/Publication.cs
@@ -2,8 +2,6 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Mnema.Models.DTOs.Content;
-using Mnema.Models.External;
-using Mnema.Models.Publication;
 
 namespace Mnema.API.Content;
 
@@ -20,17 +18,4 @@ public interface IPublication : IContent, IAsyncDisposable
     Task<MessageDto> ProcessMessage(MessageDto message);
     Task LoadMetadataAsync(CancellationTokenSource source);
     Task DownloadContentAsync(CancellationTokenSource source);
-}
-
-public class OnDiskContent: IHasPositionMarkers
-{
-    public string SeriesName { get; set; }
-    public string Path { get; set; }
-    public string FileName { get; set; }
-    public string? Chapter { get; set; }
-    public string? Volume { get; set; }
-    public ComicInfo? ComicInfo { get; set; }
-
-    public string VolumeMarker => Volume ?? string.Empty;
-    public string ChapterMarker => Chapter ?? string.Empty;
 }
